+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.3.3
+- fix pagination indicator bug (while infinite: true)
+- fix possible deadlock if more than one item is visible
+
 ### v1.3.2
 - fix event names (mouseenter, mouseleave)